refactor: refactor campaigns resolvers to have each campaign in a different file
