+++ api/src/resolvers/campaigns/equipment-campaign-cypher.ts
